requirements.txt
pybtex
aiohttp
bibtexparser
ruamel.yaml
tqdm_joblib
PyMuPDF
fitz
dask
pdf2image
pytesseract
